refactor/cart-section
